refactor: use model type instead of view name in render logic
This simplifies the code a lot and makes it more easy to extract
ipyvue.
