--- js/src/VuetifyTemplateView.js
@@ -1,8 +1,0 @@
-import { VuetifyBaseView } from './VuetifyBaseView';
-import { vueTemplateRender } from './VueTemplateRenderer';
-
-export class VuetifyTemplateView extends VuetifyBaseView {
-    vueRender(createElement) {
-        return vueTemplateRender(createElement, this.model, this);
-    }
-}
