+++ js/src/index.js
@@ -13,6 +13,5 @@ export { VuetifyView } from './VuetifyView';
 export * from './generated';
 export { HtmlModel } from './Html';
 export { VuetifyTemplateModel } from './VuetifyTemplate';
-export { VuetifyTemplateView } from './VuetifyTemplateView';
 
 export const { version } = require('../package.json'); // eslint-disable-line global-require
