+++ js/src/nodepsEmbed.js
@@ -3,6 +3,5 @@ export { VuetifyView } from './VuetifyView';
 export * from './generated';
 export { HtmlModel } from './Html';
 export { VuetifyTemplateModel } from './VuetifyTemplate';
-export { VuetifyTemplateView } from './VuetifyTemplateView';
 
 export const { version } = require('../package.json'); // eslint-disable-line global-require
